+++ src/main/java/com/rysia/conferencedemo/controllers/DiscountCodesController.java
@@ -6,9 +6,12 @@ import io.swagger.annotations.ApiOperation;
 import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
+import javax.validation.Valid;
 import java.util.List;
+import java.util.Optional;
 
 @CrossOrigin(origins = "*")
 @RestController
@@ -20,36 +23,51 @@ public class DiscountCodesController {
 
     @ApiOperation(value = "LIST ALL DISCOUNT CODES")
     @GetMapping("/codes")
-    public List<DiscountCode> list() {
-        return this.discountCodeRepository.findAll();
+    public ResponseEntity<List<DiscountCode>> list() {
+        List<DiscountCode> discountCodes = this.discountCodeRepository.findAll();
+
+        return discountCodes.isEmpty() ? new ResponseEntity<>(HttpStatus.NOT_FOUND) :
+                new ResponseEntity<List<DiscountCode>>(discountCodes, HttpStatus.OK);
     }
 
     @ApiOperation(value = "GET A UNIQUE DISCOUNT CODE")
     @GetMapping("/code/{id}")
-    public DiscountCode get(@PathVariable Long id) {
-        return this.discountCodeRepository.getOne(id);
+    public ResponseEntity<DiscountCode> get(@PathVariable(value = "id") Long id) {
+        Optional<DiscountCode> optionalDiscountCode = this.discountCodeRepository.findById(id);
+        return !optionalDiscountCode.isPresent() ? new ResponseEntity<>(HttpStatus.NOT_FOUND) :
+                new ResponseEntity<DiscountCode>(optionalDiscountCode.get(), HttpStatus.OK);
     }
 
     @ApiOperation(value = "CREATE A DISCOUNT CODE")
     @PostMapping("/code")
-    @ResponseStatus(HttpStatus.CREATED)
-    public DiscountCode create(@RequestBody final DiscountCode discountCode) {
-        return this.discountCodeRepository.saveAndFlush(discountCode);
+    public ResponseEntity<DiscountCode> create(@RequestBody @Valid final DiscountCode discountCode) {
+        return new ResponseEntity<DiscountCode>(this.discountCodeRepository.saveAndFlush(discountCode),
+                HttpStatus.CREATED);
     }
 
     @ApiOperation(value = "UPDATE A DISCOUNT CODE")
     @RequestMapping(value = "/code/{id}", method = RequestMethod.PUT)
-    @ResponseStatus(HttpStatus.OK)
-    public DiscountCode update(@PathVariable Long id, @RequestBody DiscountCode discountCode) {
-        DiscountCode existingDiscountCode = this.discountCodeRepository.getOne(id);
-        BeanUtils.copyProperties(discountCode, existingDiscountCode, "id");
-        return this.discountCodeRepository.saveAndFlush(existingDiscountCode);
+    public ResponseEntity<DiscountCode> update(@PathVariable(value = "id") Long id, @RequestBody @Valid DiscountCode discountCode) {
+        Optional<DiscountCode> optionalDiscountCode = this.discountCodeRepository.findById(id);
+        boolean existsDicount = optionalDiscountCode.isPresent();
+        DiscountCode existingDiscountCode = new DiscountCode();
+
+        if (existsDicount) {
+            existingDiscountCode = optionalDiscountCode.get();
+            BeanUtils.copyProperties(discountCode, existingDiscountCode, "id");
+        }
+
+        return existsDicount ? new ResponseEntity<DiscountCode>(this.discountCodeRepository.saveAndFlush(existingDiscountCode),
+                HttpStatus.ACCEPTED) : new ResponseEntity<>(HttpStatus.NOT_FOUND);
     }
 
     @ApiOperation(value = "DELETE A DISCOUNT CODE")
     @RequestMapping(value = "/code/{id}", method = RequestMethod.DELETE)
-    @ResponseStatus(HttpStatus.OK)
-    public void delete(@PathVariable Long id) {
-        this.discountCodeRepository.deleteById(id);
+    public ResponseEntity<?> delete(@PathVariable(value = "id") Long id) {
+        if(Optional.ofNullable(this.discountCodeRepository.findById(id)).isPresent()){
+            this.discountCodeRepository.deleteById(id);
+            return new ResponseEntity<>(HttpStatus.OK);
+        }
+        return new ResponseEntity<>(HttpStatus.NOT_FOUND);
     }
 }
